+++ src/defaults.lua
@@ -3,6 +3,7 @@ local L = LibStub("AceLocale-3.0"):GetLocale("SilentRotate")
 function SilentRotate:LoadDefaults()
 	self.defaults = {
 	    profile = {
+			-- Messaging
 	        enableAnnounces = true,
 	        channelType = "YELL",
 	        rotationReportChannelType = "RAID",
@@ -11,11 +12,28 @@ function SilentRotate:LoadDefaults()
 	        announceFailMessage = L["DEFAULT_FAIL_ANNOUNCE_MESSAGE"],
 			whisperFailMessage = L["DEFAULT_FAIL_WHISPER_MESSAGE"],
 	        announceLoathebMessage = L["DEFAULT_LOATHEB_ANNOUNCE_MESSAGE"],
-			lock = false,
-			hideNotInRaid = false,
+
+			-- Names
+			useClassColor = true,
+			useNameOutline = false,
+			appendGroup = false,
+			groupSuffix = L["DEFAULT_GROUP_SUFFIX_MESSAGE"],
+			groupSuffixColor = {SilentRotate.colors.lightGray:GetRGB()},
+
+			-- Background
+			neutralBackgroundColor = {SilentRotate.colors.lightGray:GetRGB()},
+			activeBackgroundColor  = {SilentRotate.colors.purple:GetRGB()},
+			deadBackgroundColor    = {SilentRotate.colors.red:GetRGB()},
+			offlineBackgroundColor = {SilentRotate.colors.darkGray:GetRGB()},
+
+			-- Sounds
 			enableNextToTranqSound = true,
 			enableTranqNowSound = true,
 			tranqNowSound = 'alarm1',
+
+			-- Miscellaneous
+			lock = false,
+			hideNotInRaid = false,
 			doNotShowWindowOnRaidJoin = false,
 			showWindowWhenTargetingBoss = false,
 	    },