+++ src/utils.lua
@@ -47,7 +47,7 @@ end
 
 -- Checks if a hunter is in a PvE raid
 function SilentRotate:isInPveRaid()
-    return IsInRaid() and not SilentRotate:isPlayerInBattleground()
+    return SilentRotate.testMode or IsInRaid() and not SilentRotate:isPlayerInBattleground()
 end
 
 function SilentRotate:getPlayerNameFont()
@@ -128,4 +128,35 @@ end
 -- Checks if the spell is the Rogue Distract
 function SilentRotate:isDistractSpell(spellName)
     return spellName == SilentRotate.constants.distract
+end
+
+-- Get a user-defined color or create it now
+function SilentRotate:getUserDefinedColor(colorName)
+
+    local color = SilentRotate.colors[colorName]
+
+    if (not color) then
+        -- Create the color based on profile
+        -- This should happen once, at start
+        local profileColorName
+        if (colorName == "groupSuffix") then
+            profileColorName = "groupSuffixColor"
+        else
+            profileColorName = (colorName or "").."BackgroundColor"
+        end
+
+        if (SilentRotate.db.profile[profileColorName]) then
+            color = CreateColor(
+                SilentRotate.db.profile[profileColorName][1],
+                SilentRotate.db.profile[profileColorName][2],
+                SilentRotate.db.profile[profileColorName][3]
+            )
+        else
+            print("[SilentRotate] Unknown color constant "..(colorName or "''"))
+        end
+
+        SilentRotate.colors[colorName] = color
+    end
+
+    return color
 end